# packages/types/openapi/cetem-qc-v1.yaml
openapi: 3.1.0
info:
  title: CETEM-QC API
  version: 1.0.0
  description: Versioned public HTTP contract. Internal identifiers and error codes are English.
servers:
  - url: /api/v1
paths:
...
  /history:
    get:
      operationId: listHistory
      summary: List the completed controls the authenticated user may consult (read only)
      description: >-
        Responsable (own team) or Employé (currently assigned). A control is completed when it has an official report.
        Newest designation first. Nothing is written. Request query parameters are not supported.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Completed controls (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryListResponse'
        '400': { description: Unsupported query parameters (VALIDATION_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: History could not be loaded (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /history/{taskId}:
    get:
      operationId: getHistoryRecord
      summary: Read one completed control (read only)
      description: >-
        Responsable (own team) or Employé (currently assigned). Returns the stored evidence, insights with decisions,
        confirmed summary, human conformity decision, official report and lineage. Nothing is written. A malformed,
        unknown or unauthorized task, and a task without accepted evidence or official report, all get the same 404.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The completed control (Cache-Control no-store)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryRecordResponse'
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown or unauthorized task, or not completed (TASK_NOT_FOUND); the cases are not told apart, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Stored evidence inconsistent or not loadable (INTERNAL_ERROR), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
  /history/{taskId}/official-report/file:
    get:
      operationId: downloadOfficialReport
      summary: Download the official report file of a completed control
      description: >-
        Responsable (own team) or Employé (currently assigned). Serves the stored bytes of the official candidate only,
        after checking their SHA-256. The response is an attachment with Cache-Control no-store and
        X-Content-Type-Options nosniff. No URL or storage key is returned.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The official Word document or PDF
          headers:
            Content-Disposition: { schema: { type: string } }
            Content-Length: { schema: { type: integer } }
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: { type: string, format: binary }
            application/pdf:
              schema: { type: string, format: binary }
        '401': { description: Session rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '404': { description: Malformed, unknown or unauthorized task, or not completed (TASK_NOT_FOUND), content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '409': { description: The official PDF is no longer ready (REPORT_FILE_NOT_READY); no bytes are served, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '500': { description: Stored bytes missing or altered (INTERNAL_ERROR); no bytes are served, content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
...
    ManualInsight:
      type: object
      additionalProperties: false
      required: [id, text, justification, sourceType, createdAt, author]
      properties:
        id: { type: string, format: uuid }
        text: { type: string, minLength: 1, maxLength: 1000 }
        justification: { type: [string, 'null'], minLength: 1, maxLength: 1000 }
        sourceType: { type: string, enum: [manual] }
        createdAt: { type: string, format: date-time }
        author:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
...
    ConfirmedSummary:
      type: object
      additionalProperties: false
      required: [id, version, text, confirmedAt, confirmedBy, summaryInputSetId, initialDraft]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        text: { type: string, minLength: 1, maxLength: 5000 }
        confirmedAt: { type: string, format: date-time }
        confirmedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        summaryInputSetId: { type: string, pattern: '^[0-9a-f]{64}$' }
        initialDraft:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [id, text, provider, model, requestedAt, summaryInputSetId]
              properties:
                id: { type: string, format: uuid }
                text: { type: string, minLength: 1 }
                provider: { type: string, minLength: 1 }
                model: { type: string, minLength: 1 }
                requestedAt: { type: string, format: date-time }
                summaryInputSetId: { type: string, pattern: '^[0-9a-f]{64}$' }
...
    ConformityOutcome:
      type: string
      enum: [machine-conforme, machine-non-conforme]
...
    ConformityDecision:
      type: object
      additionalProperties: false
      required: [id, outcome, decidedAt, decidedBy, summaryId, summaryVersion]
      properties:
        id: { type: string, format: uuid }
        outcome:
          $ref: '#/components/schemas/ConformityOutcome'
        decidedAt: { type: string, format: date-time }
        decidedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        summaryId: { type: string, format: uuid }
        summaryVersion: { type: integer, minimum: 1 }
    ConformityHistoryItem:
      type: object
      additionalProperties: false
      required: [id, outcome, decidedAt, decidedBy, summaryId, summaryVersion, invalidatedAt]
      properties:
        id: { type: string, format: uuid }
        outcome:
          $ref: '#/components/schemas/ConformityOutcome'
        decidedAt: { type: string, format: date-time }
        decidedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        summaryId: { type: string, format: uuid }
        summaryVersion: { type: integer, minimum: 1 }
        invalidatedAt:
          type: [string, 'null']
          format: date-time
...
    OfficialReport:
      type: object
      description: >-
        The single official report of a control, tied to the confirmed summary and the human conformity decision. No storage key.
      additionalProperties: false
      required: [id, candidateId, origin, designatedAt, designatedBy, taskId, auditId, auditRevision, submissionId, bindings, template, source, file]
      properties:
        id: { type: string, format: uuid }
        candidateId: { type: string, format: uuid }
        origin:
          type: string
          enum: [generated-word, uploaded-pdf]
        designatedAt: { type: string, format: date-time }
        designatedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        taskId: { type: string, format: uuid }
        auditId: { type: string, format: uuid }
        auditRevision: { type: integer, minimum: 1 }
        submissionId: { type: string, format: uuid }
        bindings:
          type: object
          additionalProperties: false
          required: [summaryId, summaryVersion, conformityDecisionId, conformityOutcome]
          properties:
            summaryId: { type: string, format: uuid }
            summaryVersion: { type: integer, minimum: 1 }
            conformityDecisionId: { type: string, format: uuid }
            conformityOutcome:
              $ref: '#/components/schemas/ConformityOutcome'
        template:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [id, version]
              properties:
                id: { type: string }
                version: { type: string }
        source:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [fileId]
              properties:
                fileId: { type: string, format: uuid }
        file:
          type: object
          additionalProperties: false
          required: [name, byteSize, sha256]
          properties:
            name: { type: string }
            byteSize: { type: integer, minimum: 0 }
            sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
    HistoryLineageRelation:
      type: object
      description: >-
        A replacement or recovery relation of a completed control. The related ids are present only when the viewer is
        authorized for the related task; otherwise `accessible` is false and every other field is null.
      additionalProperties: false
      required: [relation, accessible, taskId, auditId, completed]
      properties:
        relation:
          type: string
          enum: [replacement-of, replaced-by, recovery-source, recovery-successor]
        accessible: { type: boolean }
        taskId: { type: [string, 'null'], format: uuid }
        auditId: { type: [string, 'null'], format: uuid }
        completed: { type: [boolean, 'null'] }
    HistoryListItem:
      type: object
      additionalProperties: false
      required: [taskId, type, establishment, service, assignee, auditId, auditRevision, acceptedAt, designatedAt, conformityOutcome, reportOrigin, lineage]
      properties:
        taskId: { type: string, format: uuid }
        type: { type: string, enum: [graphie_mobile] }
        establishment: { type: string }
        service: { type: string }
        assignee: { type: string }
        auditId: { type: string, format: uuid }
        auditRevision: { type: integer, minimum: 1 }
        acceptedAt: { type: string, format: date-time }
        designatedAt: { type: string, format: date-time }
        conformityOutcome:
          $ref: '#/components/schemas/ConformityOutcome'
        reportOrigin:
          type: string
          enum: [generated-word, uploaded-pdf]
        lineage:
          type: array
          items:
            $ref: '#/components/schemas/HistoryLineageRelation'
    HistoryListResponse:
      type: object
      additionalProperties: false
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/HistoryListItem'
    HistoryRecordResponse:
      type: object
      description: >-
        A completed control exactly as stored: the accepted evidence of the 9.1 review plus the official report and the
        lineage relations visible to the viewer. A completed control always has a confirmed summary and a conformity decision.
      additionalProperties: false
      required: [task, submission, identity, values, results, insights, insightDecisions, manualInsights, summary, summaryVersion, summaryHistory, conformityDecision, conformityHistory, officialReport, lineage]
      properties:
        task:
          type: object
          additionalProperties: false
          required: [id, establishment, service, assignee]
          properties:
            id: { type: string, format: uuid }
            establishment: { type: string }
            service: { type: string }
            assignee: { type: string }
        submission:
          type: object
          additionalProperties: false
          required: [submissionId, auditId, revision, submittedBy, acceptedAt]
          properties:
            submissionId: { type: string, format: uuid }
            auditId: { type: string, format: uuid }
            revision: { type: integer, minimum: 1 }
            submittedBy:
              type: object
              additionalProperties: false
              required: [id, displayName]
              properties:
                id: { type: string, format: uuid }
                displayName: { type: string }
            acceptedAt: { type: string, format: date-time }
        identity:
          type: object
          additionalProperties: false
          required: [catalogueId, catalogueVersion, schemaVersion, ruleId, ruleVersion]
          properties:
            catalogueId: { type: string }
            catalogueVersion: { type: string }
            schemaVersion: { type: integer }
            ruleId: { type: string }
            ruleVersion: { type: string }
        values:
          type: object
          description: The stored payload values, strings unchanged.
          additionalProperties: { type: string }
        results:
          type: object
          description: The server calculation snapshot as stored, one entry per paper test. The shape is owned by packages/domain.
          required: [voltageAccuracy, voltageRepeatability, outputRepeatability, outputLinearity, lightFieldCorrespondence]
          additionalProperties: false
          properties:
            voltageAccuracy: { type: object, additionalProperties: true }
            voltageRepeatability: { type: object, additionalProperties: true }
            outputRepeatability: { type: object, additionalProperties: true }
            outputLinearity: { type: object, additionalProperties: true }
            lightFieldCorrespondence: { type: object, additionalProperties: true }
        insights:
          $ref: '#/components/schemas/InsightProposalSet'
        insightDecisions:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceInsightDecision'
        manualInsights:
          type: array
          items:
            $ref: '#/components/schemas/ManualInsight'
        summary:
          $ref: '#/components/schemas/ConfirmedSummary'
        summaryVersion:
          type: object
          additionalProperties: false
          required: [number, state]
          properties:
            number: { type: integer, minimum: 1 }
            state: { type: string, enum: [confirmed, open] }
        summaryHistory:
          type: array
          items:
            $ref: '#/components/schemas/SummaryHistoryItem'
        conformityDecision:
          $ref: '#/components/schemas/ConformityDecision'
        conformityHistory:
          type: array
          items:
            $ref: '#/components/schemas/ConformityHistoryItem'
        officialReport:
          $ref: '#/components/schemas/OfficialReport'
        lineage:
          type: array
          items:
            $ref: '#/components/schemas/HistoryLineageRelation'
...
    SummaryHistoryItem:
      type: object
      additionalProperties: false
      required: [version, text, confirmedAt, confirmedBy, reopenedAt, reopenedBy, initialDraft]
      properties:
        version: { type: integer, minimum: 1 }
        text: { type: string, minLength: 1, maxLength: 5000 }
        confirmedAt: { type: string, format: date-time }
        confirmedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        reopenedAt: { type: string, format: date-time }
        reopenedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        initialDraft:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [id, text, provider, model, requestedAt, summaryInputSetId]
              properties:
                id: { type: string, format: uuid }
                text: { type: string, minLength: 1 }
                provider: { type: string, minLength: 1 }
                model: { type: string, minLength: 1 }
                requestedAt: { type: string, format: date-time }
                summaryInputSetId: { type: string, pattern: '^[0-9a-f]{64}$' }
...
    EvidenceInsightDecision:
      type: object
      additionalProperties: false
      required: [proposalId, decision, decidedAt, decidedBy, registryVersion, ruleId, ruleVersion]
      properties:
        proposalId: { type: string, minLength: 1 }
        decision: { type: string, enum: [retained, discarded] }
        decidedAt: { type: string, format: date-time }
        decidedBy:
          type: object
          additionalProperties: false
          required: [id, displayName]
          properties:
            id: { type: string }
            displayName: { type: string }
        registryVersion: { type: string, minLength: 1 }
        ruleId: { type: string, minLength: 1 }
        ruleVersion: { type: integer, minimum: 1 }
    InsightSourceKey:
      type: object
      additionalProperties: false
      required: [kind, key]
      properties:
        kind: { type: string, enum: [field, result] }
        key: { type: string, minLength: 1 }
    InsightProposal:
      type: object
      additionalProperties: false
      required: [proposalId, ruleId, ruleVersion, approvalReference, registryVersion, submissionId, sourceKeys, statement, origin]
      properties:
        proposalId: { type: string, minLength: 1 }
        ruleId: { type: string, minLength: 1 }
        ruleVersion: { type: integer, minimum: 1 }
        approvalReference: { type: string, minLength: 1 }
        registryVersion: { type: string, minLength: 1 }
        submissionId: { type: string, format: uuid }
        sourceKeys:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/InsightSourceKey'
        statement: { type: string, minLength: 1 }
        origin: { type: string, enum: [deterministic] }
    InsightProposalSet:
      description: >-
        Factual proposals from CETEM-approved insight rules. unavailable means no approved rule exists; available with an
        empty list means rules exist and none fired.
      oneOf:
        - type: object
          additionalProperties: false
          required: [status, reason, registryVersion, proposals]
          properties:
            status: { type: string, enum: [unavailable] }
            reason: { type: string, enum: [no-approved-rules] }
            registryVersion: { type: string, minLength: 1 }
            proposals:
              type: array
              maxItems: 0
              items:
                $ref: '#/components/schemas/InsightProposal'
        - type: object
          additionalProperties: false
          required: [status, registryVersion, proposals]
          properties:
            status: { type: string, enum: [available] }
            registryVersion: { type: string, minLength: 1 }
            proposals:
              type: array
              items:
                $ref: '#/components/schemas/InsightProposal'
...
    ApiError:
      type: object
      description: >-
        Error envelope. Codes include VALIDATION_ERROR, AUTHENTICATION_FAILED, FORBIDDEN, TASK_NOT_FOUND,
        PAYLOAD_TOO_LARGE, IDEMPOTENCY_KEY_REUSED, AUDIT_NOT_ACCEPTED, REPLACEMENT_ALREADY_EXISTS,
        TASK_ASSIGNEE_UNAVAILABLE, AI_UNAVAILABLE, SUMMARY_ALREADY_CONFIRMED, SUMMARY_CONFIRMED, SUMMARY_NOT_CONFIRMED, SUMMARY_DESIGNATED, CONFORMITY_ALREADY_DECIDED, CONFORMITY_NOT_DECIDED, REPORT_GENERATION_FAILED,
        REPORT_INPUTS_CHANGED, REPORT_ATTEMPT_CONFLICT, REPORT_FILE_NOT_READY, REPORT_FILE_ALREADY_ATTACHED, REPORT_CANDIDATE_NOT_READY, REPORT_CANDIDATE_OUTDATED, REPORT_ALREADY_OFFICIAL, REPORT_OFFICIAL_DESIGNATED, FILE_TOO_LARGE, UNSUPPORTED_FILE_TYPE, FILE_STORAGE_FAILED,
        FILE_ATTEMPT_CONFLICT, FILE_NOT_RESCANNABLE and INTERNAL_ERROR.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                required: [path, message]
                properties:
                  path:
                    type: string
                  message:
                    type: string